fix: echo terminal not working in docs modal
