+++ apps/web/src/echo-terminal.d.ts
@@ -1,0 +1,12 @@
+import { DOMAttributes } from 'react';
+
+type CustomElement<T> = Partial<T & DOMAttributes<T> & { children: any }>;
+
+declare global {
+  // eslint-disable-next-line @typescript-eslint/no-namespace
+  namespace JSX {
+    interface IntrinsicElements {
+      ['nv-echo-terminal']: CustomElement<any>;
+    }
+  }
+}
